Moved `Seq` to /src/ser/seq.rs
